+++ crypto/txsigner_london_test.go
@@ -1,0 +1,85 @@
+package crypto
+
+import (
+	"math/big"
+	"testing"
+
+	"github.com/0xPolygon/polygon-edge/types"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLondonSigner_Sender(t *testing.T) {
+	t.Parallel()
+
+	toAddress := types.StringToAddress("1")
+
+	testTable := []struct {
+		name    string
+		chainID *big.Int
+	}{
+		{
+			"mainnet",
+			big.NewInt(1),
+		},
+		{
+			"expanse mainnet",
+			big.NewInt(2),
+		},
+		{
+			"ropsten",
+			big.NewInt(3),
+		},
+		{
+			"rinkeby",
+			big.NewInt(4),
+		},
+		{
+			"goerli",
+			big.NewInt(5),
+		},
+		{
+			"kovan",
+			big.NewInt(42),
+		},
+		{
+			"geth private",
+			big.NewInt(1337),
+		},
+		{
+			"mega large",
+			big.NewInt(0).Exp(big.NewInt(2), big.NewInt(20), nil), // 2**20
+		},
+	}
+
+	for _, testCase := range testTable {
+		testCase := testCase
+		t.Run(testCase.name, func(t *testing.T) {
+			t.Parallel()
+
+			key, keyGenError := GenerateECDSAKey()
+			if keyGenError != nil {
+				t.Fatalf("Unable to generate key")
+			}
+
+			txn := &types.Transaction{
+				To:       &toAddress,
+				Value:    big.NewInt(1),
+				GasPrice: big.NewInt(0),
+			}
+
+			signer := NewLondonSigner(testCase.chainID.Uint64())
+
+			signedTx, signErr := signer.SignTx(txn, key)
+			if signErr != nil {
+				t.Fatalf("Unable to sign transaction")
+			}
+
+			recoveredSender, recoverErr := signer.Sender(signedTx)
+			if recoverErr != nil {
+				t.Fatalf("Unable to recover sender")
+			}
+
+			assert.Equal(t, recoveredSender.String(), PubKeyToAddress(&key.PublicKey).String())
+		})
+	}
+}
